+++ README.md
@@ -4,6 +4,6 @@
 
 ## วิธีเพิ่มกฏ
 
-- ไปที่ [src/App.svelte](./src/App.svelte)
+- ไปที่ [src/rules.ts](./src/rules.ts)
 - เพิ่มฟังก์ชั่นที่แปลง String -> String
 - ส่งมาเป็น Pull Request
